requirements.txt
pandas==3.0.1
prefect==3.6.20
sqlalchemy==2.0.47
requests==2.32.5
python-dotenv==1.1.0
psycopg2-binary==2.9.11
streamlit==1.44.0
plotly==6.0.1